style: forge fmt — fix all formatting to pass CI check
747 files reformatted. CI forge fmt --check will now pass.
1041 tests passing, 0 failures.

+++ contracts/perps/tokens/interfaces/IWETH.sol
@@ -4,6 +4,6 @@ pragma solidity ^0.8.31;
 
 interface IWETH {
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
+    function transfer(address to, uint256 value) external returns (bool);
+    function withdraw(uint256) external;
 }
